[clockgen] Add vendor interface
Adds a vendor interface for querying and controlling clock selection as
the Windows UAC2 driver lacks this capability.

Uses the BOS descriptor to load the WinUSB driver with the GUID
9DED1FD1-C739-4D2A-A1E5-E060342724DE.

Requires a build of TinyUSB containing the updated ISO API for the
RP2040 to function correctly. (334ac8072650e3b18278042c2c2b402db73c7359)